i2s: Don't decode incomplete samples
oldws was initialized to 1, and then watched for changes, so:

If the signal started with WS=1, it would work correctly.

If the signal started with WS=0, the first sample would be considered a
transition of WS, and the first few bits would be decoded, even though
they are not a complete word.