i2s: Ignore errors in WAV output
Since WAV files must have a constant bit depth throughout, if the bit
depth of an I2S stream changes (such as stopping one transmission and
starting a new one in the same recording), the WAV output fails with:

    srd: return struct.pack('
    struct.error: argument out of range

This then causes the entire decoder to fail, with "Decoder reported an
error" and decoding annotations stop after that point.

With this commit, errors in bit depth are ignored in the WAV output,
replacing those samples with 0, and the decoder continues.  If the I2S
stream stops and starts, the resulting WAV will have at least one zero
sample and then will continue.